+++ CleverPush/CleverPush/Source/CPChannelTag.h
@@ -1,5 +1,6 @@
 #import <Foundation/Foundation.h>
 #import "CPUtils.h"
+
 @interface CPChannelTag : NSObject
 
 #pragma mark - Class Variables