+++ CleverPush/CleverPush/Source/CPChannelTopic.h
@@ -1,6 +1,8 @@
 #import <Foundation/Foundation.h>
 #import "CPUtils.h"
+
 @interface CPChannelTopic : NSObject
+
 #pragma mark - Class Variables
 @property (readonly, nullable) NSString *id;
 @property (readonly, nullable) NSString *parentTopic;