+++ CleverPush/CleverPush/Source/CPUtils.h
@@ -23,5 +23,6 @@
 + (BOOL)newTopicAdded:(NSDictionary*)config;
 + (NSDate*)getLocalDateTimeFromUTC:(NSString*)dateString;
 + (NSInteger)secondsBetweenDate:(NSDate*)fromDateTime andDate:(NSDate*)toDateTime;
++ (NSBundle *)getAssetsBundle;
 
 @end